+++ remote-storage-managers/s3/src/main/java/org/apache/kafka/rsm/s3/files/LogFileS3Key.java
@@ -14,19 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.rsm.s3.keys;
+package org.apache.kafka.rsm.s3.files;
+
+import org.apache.kafka.common.TopicPartition;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.kafka.common.TopicPartition;
-
-public class LogFileKey extends Key {
+/**
+ * A utility class to work with S3 keys for Kafka log files.
+ */
+public class LogFileS3Key extends S3Key {
     private static final String DIRECTORY = "log";
+    private static final Pattern NAME_PATTERN = Pattern.compile("(\\d{20})-(\\d{20})-le(\\d{10})");
 
-    private static final Pattern TOPIC_PARTITION_PATTERN = Pattern.compile("([^/])+-(\\d)+?/.*");
-
-    private LogFileKey() {}
+    private LogFileS3Key() {}
 
     public static String key(TopicPartition topicPartition, long baseOffset, long lastOffset, int leaderEpoch) {
         return keyPrefixWithoutLeaderEpochNumber(topicPartition, baseOffset, lastOffset) + formalInteger(leaderEpoch);
@@ -44,14 +46,29 @@ public class LogFileKey extends Key {
         return topicPartitionDirectory(topicPartition) + DIRECTORY_SEPARATOR + DIRECTORY + DIRECTORY_SEPARATOR;
     }
 
-    public static TopicPartition topicPartition(String s3Key) {
-        Matcher matcher = TOPIC_PARTITION_PATTERN.matcher(s3Key);
-        if (!matcher.matches()) {
-            throw new IllegalArgumentException("Can't extract topic-partition from S3 key " + s3Key);
+    public static SegmentInfo parse(String s3Key) {
+        TopicPartition topicPartition = topicPartition(s3Key);
+
+        int iSep = s3Key.indexOf(DIRECTORY_SEPARATOR);
+        if (iSep == -1) {
+            throw new IllegalArgumentException("Invalid log file S3 key format: " + s3Key);
         }
 
-        String topic = matcher.group(1);
-        int partition = Integer.parseInt(matcher.group(2));
-        return new TopicPartition(topic, partition);
+        String remaining = s3Key.substring(iSep);
+        if (!remaining.startsWith(DIRECTORY_SEPARATOR + DIRECTORY + DIRECTORY_SEPARATOR)) {
+            throw new IllegalArgumentException("Invalid log file S3 key format: " + s3Key);
+        }
+
+        remaining = remaining.substring((DIRECTORY_SEPARATOR + DIRECTORY + DIRECTORY_SEPARATOR).length());
+
+        Matcher m = NAME_PATTERN.matcher(remaining);
+        if (m.matches()) {
+            long baseOffset = Long.parseLong(m.group(1));
+            long lastOffset = Long.parseLong(m.group(2));
+            int leaderEpoch = Integer.parseInt(m.group(3));
+            return new SegmentInfo(topicPartition, baseOffset, lastOffset, leaderEpoch);
+        } else {
+            throw new IllegalArgumentException("Invalid log file S3 key format: " + s3Key);
+        }
     }
 }
