+++ remote-storage-managers/s3/src/main/java/org/apache/kafka/rsm/s3/files/OffsetIndexFileKeyS3Key.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.rsm.s3.files;
+
+import org.apache.kafka.common.TopicPartition;
+
+/**
+ * A utility class to work with S3 keys for Kafka offset index files.
+ */
+public class OffsetIndexFileKeyS3Key extends S3Key {
+    private static final String DIRECTORY = "index";
+
+    private OffsetIndexFileKeyS3Key() {}
+
+    public static String key(TopicPartition topicPartition, long baseOffset, long lastOffset, int leaderEpoch) {
+        return topicPartitionDirectory(topicPartition) + DIRECTORY_SEPARATOR + DIRECTORY + DIRECTORY_SEPARATOR
+            + formatLong(baseOffset)
+            + "-"
+            + formatLong(lastOffset)
+            + "-le" + formalInteger(leaderEpoch);
+    }
+}
